requirements.txt
discord.py>=2.0.0
twstock
beautifulsoup4
requests
